+++ packages/react/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add `"sideEffects": false` for better tree-shaking.
 - Add `useLocales` to get the context without immediately converting it to the translator object.
   - `useI18n` now internally uses `useLocales`. Be careful if you have a locale named `""` (empty string), which I believe is very rare.
+- `useI18n` now memoizes the translator object.
+  You get the identical object or function in the same element until the book or the locale changes.
 
 ## 0.1.1